clean(api): clean oauth onboarding
