+++ packages/db/index.ts
@@ -15,3 +15,4 @@ neonConfig.fetchConnectionCache = true;
 
 const sql = neon("process.env.DATABASE_URL");
 export const db = drizzle(sql, { schema });
+export type dbType = typeof db;
